Phase 3.5: Add CPU TMP word register with 128-bit control words
Changes:
- Add TMP WordRegister to CPU (connected to address bus)
- Expand control word from 64-bit to 128-bit to accommodate 67 controls
- Update HDL Cpu to include tmp register with controls
- Update MicrocodeTable to use __uint128_t
- Update all encoders to support 128-bit control words
- Update MicrocodeRomStorage from <uint32_t, uint64_t> to <uint32_t, __uint128_t>
- Update component count tests to reflect new register

The TMP register provides a temporary word register for complex addressing
modes and word operations. With its 3 controls (write, read, reset), the
total control count increased from 64 to 67, necessitating the expansion
to 128-bit control words.

All 298 tests pass.

+++ microcode/include/irata2/microcode/encoder/control_encoder.h
@@ -19,8 +19,8 @@ class ControlEncoder {
  public:
   explicit ControlEncoder(const hdl::Cpu& cpu);
 
-  uint64_t Encode(const std::vector<const hdl::ControlInfo*>& controls) const;
-  std::vector<std::string> Decode(uint64_t control_word) const;
+  __uint128_t Encode(const std::vector<const hdl::ControlInfo*>& controls) const;
+  std::vector<std::string> Decode(__uint128_t control_word) const;
 
   const std::vector<std::string>& control_paths() const { return control_paths_; }
